refactoring mostly
