+++ src/app/settings/page.tsx
@@ -4,7 +4,7 @@ import { AppContext } from '@/context/context';
 import { useRouter } from 'next/navigation';
 import { useContext } from 'react';
 
-export default function () {
+export default function Settings() {
   const context = useContext(AppContext);
   if (!context) {
     throw new Error('This component requires AppProvider as a parent');
@@ -12,12 +12,12 @@ export default function () {
 
   const router = useRouter();
   return (
-    <>
+    <div>
       <div>settings</div>
       <div>{context.user?.login}</div>
       <button onClick={() => router.push(`/${context.order.join('/')}`)}>
         go back
       </button>
-    </>
+    </div>
   );
 }
