+++ src/components/header.tsx
@@ -1,17 +1,9 @@
-import {
-  Carousel,
-  Grid,
-  LogOut,
-  Plus,
-  Settings,
-  Twitch,
-} from '@/components/icons';
+import { Carousel, Grid, Plus, Twitch } from '@/components/icons';
 import { PlayerView } from '@/types/state';
 import { User } from '@/types/twitch';
-import Image from 'next/image';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-import { FormEvent, useEffect, useState } from 'react';
+import { FormEvent } from 'react';
+import UserMenu from './user-menu';
 
 const TWITCH_CLIENT_ID = process.env.NEXT_PUBLIC_TWITCH_CLIENT_ID;
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
@@ -39,17 +31,6 @@ export default function Header({
     if (channel.length <= 0) return;
     addWatching(channel);
   };
-
-  const [userMenuOpen, setUserMenuOpen] = useState<boolean>(false);
-  const toggleUserMenu = () => {
-    setUserMenuOpen(!userMenuOpen);
-  };
-
-  const pathname = usePathname();
-  useEffect(() => {
-    setUserMenuOpen(false);
-  }, [pathname]);
-
   return (
     <header className="flex h-20 grow-0 shrink-0 bg-chatpanel z-20 header-shadow justify-between">
       <div className="flex">
@@ -104,71 +85,7 @@ export default function Header({
         </form>
       </div>
       <div className="relative flex p-4">
-        {user && (
-          <>
-            <div className="w-[30px] cursor-pointer" onClick={toggleUserMenu}>
-              <Image
-                src={user.profile_image_url}
-                alt={`${user.login} profile`}
-                className="w-full max-w-full rounded-full object-cover"
-                width={30}
-                height={30}
-                unoptimized
-              />
-            </div>
-            <div
-              id="user-menu"
-              className={`${
-                userMenuOpen ? 'block' : 'hidden'
-              } absolute top-[4.5rem] right-4 z-30 flex flex-col bg-sidepanel p-4 min-w-[200px] user-menu-shadow rounded-l-xl border-r border-twbuttontext text-sm`}
-            >
-              <div className="flex items-center p-2">
-                <div className="w-[40px]">
-                  <Image
-                    src={user.profile_image_url}
-                    alt={`${user.login} profile`}
-                    className="w-full max-w-full rounded-full object-cover"
-                    width={40}
-                    height={40}
-                    unoptimized
-                  />
-                </div>
-                <span className="pl-4 font-semibold">{user.display_name}</span>
-              </div>
-              <div
-                role="separator"
-                className="border-t border-twborder mt-4 mx-2 pb-4"
-              ></div>
-              <div>
-                <Link
-                  href="/settings"
-                  className="flex p-2 rounded-lg hover:bg-twbuttonbg hover:bg-opacity-[0.48]"
-                >
-                  <div className="pr-2 h-[20px]">
-                    <Settings className="fill-twbuttontext" />
-                  </div>
-                  <span className="">Settings</span>
-                </Link>
-              </div>
-              <div
-                role="separator"
-                className="border-t border-twborder mt-4 mx-2 pb-4"
-              ></div>
-              <div>
-                <a
-                  href="https://www.twitch.tv/settings/connections"
-                  target="_blank"
-                  className="flex p-2 rounded-lg hover:bg-twbuttonbg hover:bg-opacity-[0.48]"
-                >
-                  <div className="pr-2 h-[20px]">
-                    <LogOut className="fill-twbuttontext" />
-                  </div>
-                  <span className="">Log Out</span>
-                </a>
-              </div>
-            </div>
-          </>
-        )}
+        <UserMenu user={user} />
         {!accessToken && (
           <a
             href={`https://id.twitch.tv/oauth2/authorize?response_type=token&client_id=${TWITCH_CLIENT_ID}&redirect_uri=${BASE_URL}&scope=user:read:follows`}
